sequence 修改，接口列表样式修改

+++ api/src/main/java/cn/crap/controller/admin/SettingController.java
@@ -114,7 +114,7 @@ public class SettingController extends BaseController {
     public JsonResult changeSequence(@RequestParam String id, @RequestParam String changeId) {
         Setting change = settingService.getById(changeId);
         Setting model = settingService.getById(id);
-        int modelSequence = model.getSequence();
+        long modelSequence = model.getSequence();
 
         model.setSequence(change.getSequence());
         change.setSequence(modelSequence);
