+++ api/src/main/java/cn/crap/dto/ProjectDto.java
@@ -8,7 +8,7 @@ public class ProjectDto {
 	private String id;
 	private String name;
 	private Byte status;
-	private Integer sequence;
+	private Long sequence;
 	private String remark;
 	private String userId;
 	private String userName;
@@ -51,10 +51,10 @@ public class ProjectDto {
 		return status;
 	}
 
-	public void setSequence(Integer sequence){
+	public void setSequence(Long sequence){
 		this.sequence=sequence;
 	}
-	public Integer getSequence(){
+	public Long getSequence(){
 		return sequence;
 	}
